% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachLineups.R
\name{attachLineups}
\alias{attachLineups}
\title{Extracts and attachs lineups at all events of a play-by-play data frame}
\usage{
attachLineups(pbp)
}
\arguments{
\item{pbp}{Play-by-play data frame.}
}
\value{
The original play-by-play data with 11 extra variables identifying
  the players that were on the floor at each event.
}
\description{
Extracts and attachs lineups at all events of a play-by-play data frame
}
\keyword{internal}
