% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShotchart.R
\name{construct_arc}
\alias{construct_arc}
\title{Construct the coordinates of an arc}
\usage{
construct_arc(x0, y0, r, start, stop)
}
\arguments{
\item{x0}{Center x coordinate}

\item{y0}{Center y coordinate}

\item{r}{Radius}

\item{start}{In radians from 0 to pi}

\item{stop}{In radians from 0 to pi}
}
\value{
Data frame with x, y coordinates
}
\description{
Construct the coordinates of an arc
}
\examples{
circle <- construct_arc(x0 = 0, y0 = 0, r = 1, start = 0, stop = 2 * pi)
plot(circle$x, circle$y, type = "l")
arc <- construct_arc(2, 1, 5, start = pi, stop = pi / 2)
plot(arc$x, arc$y, type = "l")
}
\keyword{internal}
