% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShotchart.R
\name{construct_court}
\alias{construct_court}
\title{Construct coordinates of basketball court}
\usage{
construct_court()
}
\value{
A list containing data frames corresponding to different parts of
  the court
}
\description{
Construct coordinates of basketball court
}
\examples{
court <- construct_court()
}
