% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurolig.R
\docType{package}
\name{eurolig}
\alias{eurolig}
\alias{eurolig-package}
\title{eurolig: A package for obtaining and analyzing data from Euroleague}
\description{
eurolig: A package for obtaining and analyzing data from Euroleague
}
