% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pbp.R
\name{extract_pbp}
\alias{extract_pbp}
\title{Extract play-by-play data}
\usage{
extract_pbp(game_code, season)
}
\arguments{
\item{game_code}{An integer specifying the game code}

\item{season}{An integer specifying the starting year of the desired season}
}
\value{
A tibble
}
\description{
\code{extract_pbp} downloads play-by-play data from Euroleague and returns
a tidy data frame.
}
