% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_team_pbp.R
\name{extract_team_pbp}
\alias{extract_team_pbp}
\title{Extract play-by-play data for a given team}
\usage{
extract_team_pbp(team, season)
}
\arguments{
\item{team}{A three letter string specifying the team code}

\item{season}{An integer specifying the starting year of the desired season}
}
\value{
A tibble
}
\description{
Downloads play-by-play data of all games that a given team has played
in the specified season
}
