% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMultiplePbp.R
\name{extractMultiplePbp}
\alias{extractMultiplePbp}
\title{Title}
\usage{
extractMultiplePbp(game_codes, season)
}
\arguments{
\item{game_codes}{Numeric vector of game codes}

\item{season}{Scalar}
}
\description{
Title
}
\keyword{internal}
