% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPbp.R
\name{extractPbp}
\alias{extractPbp}
\title{Extracts play-by-play data for a single game}
\usage{
extractPbp(game_code, season, lineups = TRUE)
}
\arguments{
\item{game_code}{Integer scalar specifying the game code}

\item{season}{Integer scalar specifying the season}

\item{lineups}{Logical scalar indicating whether to include
lineup information. Defaults to TRUE.}
}
\value{
A play-by-play data frame with the following variables:
  \describe{
    \item{season}{Starting year of the season}
    \item{game_code}{Game code}
    \item{play_number}{Identifying event number}
    \item{team_code}{Team code of involved player}
    \item{player_name}{Full player name (Last name, First name)}
    \item{play_type}{Play type. See Details.}
    \item{time_remaining}{Time remaining in the quarter}
    \item{quarter}{Quarter}
    \item{points_home}{Home team's score}
    \item{points_away}{Away team's score}
    \item{play_info}{Event description}
    \item{seconds}{Seconds elapsed since the start of the game}
    \item{home_team}{Full name of home team}
    \item{away_team}{Full name of away team}
    \item{home}{Whether player involved plays at home}
    \item{team_name}{Full team name of involved player}
    \item{last_ft}{Whether event corresponds to a last free throw in a trip to the line}
    \item{and1}{Whether event corresponds to an additional free throw}
    \item{home_player1:home_player5}{Home players on the floor}
    \item{away_player1:away_player5}{Away players on the floor}
    \item{lineups}{All 10 players on the floor}
  }
}
\description{
\code{extractPbp} returns a data frame with the play-by-play data for
the specified game.
}
\details{
Each game is uniquely identified by the game code and the season. Game codes
can be found in the included data set \code{\link{gameresults}} or using
\code{\link{extractResults}}.

Columns indicating the 10 players that were on the court at each event
of the play-by-play data are included by default.
}
\section{Warning}{

Euroleague's API documentation asks for a delay of 15 secods per request.
Take this into consideration when requesting a large number of games.
}

\examples{
extractPbp(game_code = 1, season = 2018)
}
\seealso{
Other extract functions: \code{\link{extractResults}}
}
\concept{extract functions}
