% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{extractResults}
\alias{extractResults}
\title{Extract game results from a given season}
\usage{
extractResults(season)
}
\arguments{
\item{season}{Integer indicating the season}
}
\value{
A data frame with all the games in the given season
}
\description{
Extract game results from a given season
}
