% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractShots.R
\name{extractShots}
\alias{extractShots}
\title{Extract shot location data for a single game}
\usage{
extractShots(game_code, season)
}
\arguments{
\item{game_code}{Integer scalar}

\item{season}{Integer scalar}
}
\value{
Data frame with shot location data and additional contextual
  information about each shot.
}
\description{
\code{extractShots} returns a data frame with coordinate location data on
every shot taken in the specified game.
}
