% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTeamPbp.R
\name{extractTeamPbp}
\alias{extractTeamPbp}
\title{Extract play-by-play data for a given team}
\usage{
extractTeamPbp(team, season)
}
\arguments{
\item{team}{A three letter string specifying the team code}

\item{season}{An integer specifying the starting year of the desired season}
}
\value{
A tibble
}
\description{
Downloads play-by-play data of all games that a given team has played
in the specified season
}
\section{Warning}{

This functions may take over a minute to return the large output data frame
}

