% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTeams.R
\name{extractTeams}
\alias{extractTeams}
\title{Extract info of the teams for a given season}
\usage{
extractTeams(season)
}
\arguments{
\item{season}{Integer value indicating the season}
}
\value{
A data frame with information of all the teams in the given season
}
\description{
Extract info of the teams for a given season
}
