% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixLineup.R
\name{fixLineups}
\alias{fixLineups}
\title{Correct lineups in play-by-play data}
\usage{
fixLineups(pbp)
}
\arguments{
\item{pbp}{}
}
\description{
\code{fixLineups} corrects the lineups in free throw stints so that free trhows are attributed to the lineups that were on the court when the foul occured
}
