% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gameresults}
\alias{gameresults}
\title{Euroleague game results (2001-2018)}
\format{A data frame with 4,173 rows and 13 variables:
\describe{
  \item{season}{Year the season started.}
  \item{phase}{Phase of the season: regular season (RS), top 16 (TS),
    Playoffs (PO), and Final Four (FF).}
  \item{team_home}{Full team name of the home team.}
  \item{points_home}{Points scored by the home team.}
  \item{team_away}{Full team name of the away team.}
  \item{points_away}{Points scored by the away team.}
  \item{game_code}{Game code partially identifying the game.}
  \item{date}{Date of the game.}
  \item{round_name}{The round within a phase.}
  \item{round_code}{Round code.}
  \item{game_url}{Game url.}
  \item{team_code_home}{Team code of home team.}
  \item{team_code_away}{Team code of away team.}
}}
\source{
\url{https://www.euroleague.net/}
}
\usage{
gameresults
}
\description{
Each game is uniquely idendified by the season and game code.
}
\keyword{datasets}
