% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assists.R
\name{get_assists}
\alias{get_assists}
\title{Get who assisted who from play-by-play data}
\usage{
get_assists(df, team)
}
\arguments{
\item{df}{Play-by-play data frame as produced by extract_pbp()}

\item{team}{A three letter string specifying the team code}
}
\value{
A data frame
}
\description{
Get who assisted who from play-by-play data
}
