% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineups.R
\name{get_lineups}
\alias{get_lineups}
\title{Get lineups at all events of a play-by-play data frame}
\usage{
get_lineups(pbp)
}
\arguments{
\item{pbp}{Data frame as obtained by extract_pbp()}
}
\value{
Data frame with play-by-play data and lineups attached
}
\description{
Get lineups at all events of a play-by-play data frame
}
