% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssists.R
\name{getAssists}
\alias{getAssists}
\title{Get who assisted who from play-by-play data}
\usage{
getAssists(pbp, team)
}
\arguments{
\item{pbp}{Play-by-play data frame as produced by extract_pbp()}

\item{team}{An optional three letter string specifying the team code}
}
\value{
A data frame
}
\description{
Get who assisted who from play-by-play data
}
