% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixLineups.R
\name{getFtLineup}
\alias{getFtLineup}
\title{Get the lineup on the court at the start of a free throw stint}
\usage{
getFtLineup(ft_stint)
}
\arguments{
\item{ft_stint}{A data frame}
}
\value{
A data frame with the corrected lineups
}
\description{
Get the lineup on the court at the start of a free throw stint
}
