% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOnOff.R
\name{getOnOffStats}
\alias{getOnOffStats}
\title{Get on/off statistics from play-by-play data}
\usage{
getOnOffStats(pbp, players)
}
\arguments{
\item{pbp}{A play-by-play data frame}

\item{players}{Character vector of full player names}
}
\value{
A data frame with 4 rows per game and 28 variables:
  \describe{
    \item{season}{Starting year of the season}
    \item{game_code}{Game code}
    \item{players}{Players that were on/off the court}
    \item{on}{Whether players were on the court}
    \item{type}{Factor indicating whether the stats correspond to offense/defense}
    \item{team_code}{Team code.}
    \item{home}{Whether the team was playing at home or not}
    \item{fg2a}{2PA}
    \item{fg2m}{2PM}
    \item{fg2_pct}{2P\%}
    \item{fg3a}{3PA}
    \item{fg3m}{3PM}
    \item{fg3_pct}{3P\%}
    \item{fga}{FGA}
    \item{fgm}{FGM}
    \item{fg_pct}{FG\%}
    \item{fta}{FTA}
    \item{ftm}{FTM}
    \item{ft_pct}{FT\%}
    \item{orb}{ORB}
    \item{drb}{DRB}
    \item{tov}{TOV}
    \item{ast}{AST}
    \item{stl}{STL}
    \item{cpf}{Comitted personal fouls}
    \item{blk}{BLK}
    \item{pts}{PTS}
    \item{poss}{Possessions}
  }
}
\description{
\code{getOnOffStats} returns team statistics on offense and defense
when the specified players were on the court together and when they
were on the bench.
}
\examples{
data(samplepbp)
madrid_players <- c("FERNANDEZ, RUDY", "AYON, GUSTAVO")
getOnOffStats(pbp = samplepbp, players = madrid_players)
}
