% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPossEnding.R
\name{getPbpPoss}
\alias{getPbpPoss}
\title{Get number of possessions from play-by-play data}
\usage{
getPbpPoss(pbp)
}
\arguments{
\item{pbp}{A play-by-play data frame}
}
\value{
A data frame with number of offensive and defensive possesions
}
\description{
\code{getPbpPoss} returns the number of possesions of each team from
the play-by-play data of a single game
}
