% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPbpStats.R
\name{getPbpStats}
\alias{getPbpStats}
\title{Get team statistics from play-by-play data}
\usage{
getPbpStats(pbp)
}
\arguments{
\item{pbp}{A play-by-play data frame}
}
\description{
\code{getPbpStats} returns boxscore statistics from play-by-play data from
one or more games.
}
