% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlusMinus.R
\name{getPlusMinus}
\alias{getPlusMinus}
\title{Get plus-minus for the time a player is on the court}
\usage{
getPlusMinus(pbp, players)
}
\arguments{
\item{pbp}{Play-by-play data frame}

\item{players}{Character vector with the players on the court}
}
\value{
A data frame with plus/minus information for each game
}
\description{
\code{getPlusMinus} returns a data frame with the players' plus/minus stats
for each game.
}
