% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPm.R
\name{getPm}
\alias{getPm}
\title{Get plus-minus for the time a player is on the court}
\usage{
getPm(pbp, players)
}
\arguments{
\item{pbp}{Play-by-play data frame}

\item{players}{Character vector with the players on the court}
}
\description{
Get plus-minus for the time a player is on the court
}
