% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPossEnding.R
\name{getPossEnding}
\alias{getPossEnding}
\title{Find when a pbp event ends a possession}
\usage{
getPossEnding(pbp)
}
\arguments{
\item{pbp}{Play-by-play data frame}
}
\description{
Find when a pbp event ends a possession
}
