% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSecondsElapsed.R
\name{getSecondsElapsed}
\alias{getSecondsElapsed}
\title{Get seconds elapsed since the start of the game}
\usage{
getSecondsElapsed(quarters, secs)
}
\arguments{
\item{quarters}{Integer vector indicating the quarter}

\item{secs}{Character vector indicating minute and
seconds left in the quarter (MM:SS)}
}
\description{
Get seconds elapsed since the start of the game
}
\keyword{internal}
