% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPm.R
\name{getStintStats}
\alias{getStintStats}
\title{Get stats for a single stint}
\usage{
getStintStats(pbp, players)
}
\arguments{
\item{pbp}{Play-by-play data frame}

\item{players}{Character vector with the players on the court}
}
\value{
Data frame with stats on offense and defense when players were on the court
}
\description{
\code{getStintStats} parses a play-by-play data frame and returns the statistics while the players were on the court
}
