% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPm.R
\name{getStintStats2}
\alias{getStintStats2}
\title{Get stats for a single stint}
\usage{
getStintStats2(pbp, players, team)
}
\arguments{
\item{pbp}{Play-by-play data frame}

\item{players}{Character vector with the players on the court}

\item{team}{Three letter string indicating the team code}
}
\value{
Data frame with stats on offense and defense when players were on the court
}
\description{
\code{getStintStats2} parses a play-by-play data frame and returns the statistics while the players were on the court
}
