% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTeamGameCodes.R
\name{getTeamResults}
\alias{getTeamResults}
\title{Get team results for a single season}
\usage{
getTeamResults(team, season)
}
\arguments{
\item{team}{Team code}

\item{season}{Integer scalar}
}
\description{
Get team results for a single season
}
