% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isOnCourt.R
\name{isOnCourt}
\alias{isOnCourt}
\title{Find if players were on the court}
\usage{
isOnCourt(pbp, players)
}
\arguments{
\item{pbp}{A play-by-play data frame}

\item{players}{A character vector with the nemes of the players}
}
\description{
\code{isOnCourt} finds if a player or a set of players is on the court
for each event of play-by-play data with lineup information.
}
