% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPbp.R
\name{makePointsVar}
\alias{makePointsVar}
\title{Create home and away points variables}
\usage{
makePointsVar(df)
}
\arguments{
\item{df}{A data frame of play-by-play data}
}
\value{
A data frame with home and away points for each action of pbp
}
\description{
Helper function to create points variables. Used to add these variables to output data frame in extract_pbp().
}
