% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assists.R
\name{plot_assists}
\alias{plot_assists}
\title{Plot assists network}
\usage{
plot_assists(pbp_df, team_code)
}
\arguments{
\item{pbp_df}{Data frame with play-by-play data}

\item{team}{}
}
\description{
Uses igpraph to plot a network graph showing who assisted who from play-by-play data.
}
