% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot assist heatmap}
\usage{
plot_heatmap(pbp, team)
}
\arguments{
\item{pbp}{Play-by-play data frame}

\item{team}{A three letter string specifying the team code}
}
\value{
A ggplot heatmap
}
\description{
\code{plot_heatmap} creates a heatmap indicating who assists who
}
