% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot assists network}
\usage{
plot_network(pbp_df, team, n_min)
}
\arguments{
\item{pbp_df}{Data frame with play-by-play data}

\item{team}{A three letter string specifying the team code}
}
\value{
An igraph network plot
}
\description{
Uses igpraph to plot a network graph showing who assisted who from play-by-play data.
}
