% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plotAstNetwork}
\alias{plotAstNetwork}
\title{Plot assists network}
\usage{
plotAstNetwork(pbp_df, team, n_min)
}
\arguments{
\item{pbp_df}{Data frame with play-by-play data}

\item{team}{A three letter string specifying the team code}

\item{n_min}{Integer scalar specifying the minimum number of assists to be
considered}
}
\value{
An igraph network plot
}
\description{
Uses igpraph to plot a network graph showing who assisted who from play-by-play data.
}
