% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShotchart.R
\name{plotShotchart}
\alias{plotShotchart}
\title{Plot shotcharts}
\usage{
plotShotchart(shots)
}
\arguments{
\item{shots}{}
}
\description{
\code{plotShotchart} plots where shots were taking on the court.
}
