% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requestShots.R
\name{requestShots}
\alias{requestShots}
\title{Request shot location data from Euroleague's API}
\usage{
requestShots(game_code, season)
}
\arguments{
\item{game_code}{Integer scalar}

\item{season}{Integer scalar}
}
\description{
\code{requestShots} returns the raw data frame obtained when using
Euroleague's API.
}
\keyword{internal}
