% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samplepbp}
\alias{samplepbp}
\title{Euroleague 2018/2019 Final Four play-by-play dataset}
\format{A play-by-play data frame with 2,121 rows and 29 variables:
  \describe{
    \item{season}{Starting year of the season}
    \item{game_code}{Game code}
    \item{play_number}{Identifying event number}
    \item{team_code}{Team code of involved player}
    \item{player_name}{Full player name (Last name, First name)}
    \item{play_type}{Play type. See Details.}
    \item{time_remaining}{Time remaining in the quarter}
    \item{quarter}{Quarter}
    \item{points_home}{Home team's score}
    \item{points_away}{Away team's score}
    \item{play_info}{Event description}
    \item{seconds}{Seconds elapsed since the start of the game}
    \item{home_team}{Full name of home team}
    \item{away_team}{Full name of away team}
    \item{home}{Whether player involved plays at home}
    \item{team_name}{Full team name of involved player}
    \item{last_ft}{Whether event corresponds to a last free throw in a trip to the line}
    \item{and1}{Whether event corresponds to an additional free throw}
    \item{home_player1:home_player5}{Home players on the floor}
    \item{away_player1:away_player5}{Away players on the floor}
    \item{lineups}{All 10 players on the floor}
  }}
\source{
\url{https://www.euroleague.net/}
}
\usage{
samplepbp
}
\description{
Euroleague 2018/2019 Final Four play-by-play dataset
}
\keyword{datasets}
