% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_pbp.R
\name{scrape_pbp}
\alias{scrape_pbp}
\title{Scrape data from Eurleague website}
\usage{
scrape_pbp(game_code, season)
}
\arguments{
\item{game_code}{An integer as specified in the game url}

\item{season}{The starting year of the desired season}
}
\value{
A data frame of raw data returned by the API
}
\description{
Scrape data from Eurleague website
}
