% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_team_results.R
\name{scrape_team_results}
\alias{scrape_team_results}
\title{Extract results for a team}
\usage{
scrape_team_results(team, season)
}
\arguments{
\item{team}{A three letter string specifying the team code}

\item{season}{An integer specifying the starting year of the desired season}
}
\value{
A tibble
}
\description{
Get results of the games of a given team and season
}
