% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teaminfo}
\alias{teaminfo}
\title{Euroleague teams (2001-2019)}
\format{A data frame with 434 rows and 5 variables:
\describe{
  \item{season}{Year the season started.}
  \item{team_code}{Team code.}
  \item{team_name}{Full team name.}
  \item{team_link}{Team URL in Euroleague's site.}
  \item{team_img}{Link to image file of team logo.}
}}
\source{
\url{https://www.euroleague.net/}
}
\usage{
teaminfo
}
\description{
Euroleague teams (2001-2019)
}
\keyword{datasets}
