% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichLastFt.R
\name{whichLastFt}
\alias{whichLastFt}
\title{Find all last free throws in play-by-play data}
\usage{
whichLastFt(pbp)
}
\arguments{
\item{pbp}{Play-by-play data frame}
}
\value{
Integer vector with the indexes of the last free throws
}
\description{
Find all last free throws in play-by-play data
}
