% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichLastFt.R
\name{whichLastFtInStint}
\alias{whichLastFtInStint}
\title{Find last free throw of a single trip to the foul line}
\usage{
whichLastFtInStint(ft_stint)
}
\arguments{
\item{ft_stint}{Data frame with all events in a single trip to the line}
}
\value{
Integer scalar indicating the play_number of the last free throw in the stint
}
\description{
Find last free throw of a single trip to the foul line
}
\keyword{internal}
