% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachLineups.R
\name{attachLineups}
\alias{attachLineups}
\title{Extracts and attachs lineups at all events of a play-by-play data frame}
\usage{
attachLineups(pbp)
}
\arguments{
\item{pbp}{Play-by-play data frame.}
}
\value{
Data frame with play-by-play data
}
\description{
Extracts and attachs lineups at all events of a play-by-play data frame
}
