% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineups.R
\name{attachLineups}
\alias{attachLineups}
\title{Attach lineups at all events of a play-by-play data frame}
\usage{
attachLineups(pbp)
}
\arguments{
\item{pbp}{Data frame as obtained by extract_pbp()}
}
\value{
Data frame with play-by-play data
}
\description{
Attach lineups at all events of a play-by-play data frame
}
