% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pbp.R
\name{extract_pbp}
\alias{extract_pbp}
\title{Extract play-by-play data for Euroleague games}
\usage{
extract_pbp(game_code, season_code)
}
\arguments{
\item{game_code}{An integer as specified in the game url}

\item{season_code}{A string as specified in the game url}
}
\value{
A data frame
}
\description{
Extract play-by-play data for Euroleague games
}
