% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pbp.R
\name{extract_pbp}
\alias{extract_pbp}
\title{Extract play-by-play data for Euroleague games}
\usage{
extract_pbp(game_code, season)
}
\arguments{
\item{game_code}{An integer as specified in the game url}

\item{season}{The starting year of the desired season}
}
\value{
A data frame
}
\description{
Extract play-by-play data for Euroleague games
}
