% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPbp.R
\name{extractPbp}
\alias{extractPbp}
\title{Extracts play-by-play data for a single game}
\usage{
extractPbp(game_code, season, lineups = TRUE)
}
\arguments{
\item{game_code}{Integer scalar specifying the game code}

\item{season}{Integer scalar specifying the season}

\item{lineups}{Logical scalar indicating whether to include
lineup information. Defaults to TRUE.}
}
\value{
A play-by-play data frame with the following variables:
  \describe{
    \item{season}{Starting year of the season.}
    \item{game_code}{Game code.}
  }
}
\description{
\code{extractPbp} returns a data frame with the play-by-play data for
the specified game.
}
\details{
Each game is uniquely identified by the game code and the season. Game codes
can be found in the included data set \code{\link{gameresults}} or using
\code{\link{extractResults}}.

Columns indicating the 10 players that were on the court at each event
of the play-by-play data are included by default.
}
\section{Warning}{

Euroleague's API documentation asks for a delay of 15 secods per request.
Take this into consideration when requesting a large number of games.
}

