% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults.R
\name{extractResults}
\alias{extractResults}
\title{Extract game results from a given season}
\usage{
extractResults(season)
}
\arguments{
\item{season}{Integer indicating the season}
}
\value{
A data frame with all the games in the given season
}
\description{
\code{extractResults} returns a data frame with all game results for
a single season.
}
\details{
\code{extractResults} is useful for the current season (2019).
For past seasons check the included data set \code{gameresults}.
}
\section{Note}{

\code{extractResults} takes considerable time since it parses all
web pages containing game results for the given season. Looping over
many seasons may take a long time.
}

