% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStintStats.R
\name{getStintStats}
\alias{getStintStats}
\title{Get stats for stints that players were on the floor}
\usage{
getStintStats(pbp, players, format_long = TRUE)
}
\arguments{
\item{pbp}{Play-by-play data frame.}

\item{players}{Character vector with the players on the court.}

\item{format_long}{Logical scalar indicating how output data frame should
be returned (defaults to TRUE).}
}
\value{
Data frame with stats on offense and defense when players were
  on the court. If \code{format_long = TRUE} there will be two rows per
  game: one for offense and one for defense. If \code{format_long = FALSE}
  offense and defense stats for each game will be on the same row.
}
\description{
\code{getStintStats} parses a play-by-play data frame and returns the
statistics while the players were on the court.
}
