% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPm.R
\name{getStintStats}
\alias{getStintStats}
\title{Get stats for a single stint}
\usage{
getStintStats(pbp, players)
}
\arguments{
\item{pbp}{Play-by-play data frame}

\item{players}{Character vector with the players on the court}
}
\value{
Data frame with stats on offense and defense
}
\description{
Get stats for a single stint
}
