% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTeamGameCodes.R
\name{getTeamGameCodes}
\alias{getTeamGameCodes}
\title{Get game codes for a team}
\usage{
getTeamGameCodes(team, season)
}
\arguments{
\item{team}{Team code. Each team has a three letter code that uniquely
identifies it. See \code{\link{teaminfo}}.}

\item{season}{Integer scalar identifying the season.}
}
\description{
\code{getGameCodes} returns the game codes for all the games played by
the specified team in a season.
}
