% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShotchart.R
\name{plotShotchart}
\alias{plotShotchart}
\title{Plot shotcharts}
\usage{
plotShotchart(shots)
}
\arguments{
\item{shots}{A shot location data frame.}
}
\description{
\code{plotShotchart} plots where shots were taken on the court.
}
